requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
deepgram-sdk==3.2.7
python-dotenv==1.0.0
python-multipart==0.0.6
google-generativeai==0.3.2
pydantic==2.5.0 